// src/common/enums/event.enum.ts
export enum EventType {
  FREE = 'Gratuito',
  PAID = 'Pago',
}

export enum EventStatus {
  ACTIVE = 'Ativo',
  INACTIVE = 'Inativo',
  CANCELLED = 'Cancelado',
  COMPLETED = 'Completo',
}

export enum EventCategory {
  EVENT = 'Evento',
  MEETING = 'Reunião',
  CONFERENCE = 'Conferência',
  WORKSHOP = 'Seminário',
  CONGRESS = 'Congresso',
}